# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/token":{"post":{"summary":"Login For Access Token","operationId":"login_for_access_token_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/employees/":{"post":{"tags":["Employees"],"summary":"Create Employee","operationId":"create_employee_employees__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Employees"],"summary":"Read Employees","operationId":"read_employees_employees__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeOut"},"title":"Response Read Employees Employees  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/employees/{employee_id}":{"get":{"tags":["Employees"],"summary":"Read Employee","operationId":"read_employee_employees__employee_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Employees"],"summary":"Update Employee","operationId":"update_employee_employees__employee_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Employees"],"summary":"Delete Employee","operationId":"delete_employee_employees__employee_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/attendance/attendance/log":{"post":{"tags":["Attendance"],"summary":"Log Attendance","operationId":"log_attendance_attendance_attendance_log_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/attendance/":{"get":{"tags":["Attendance"],"summary":"Read Attendance","operationId":"read_attendance_attendance__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttendanceOut"},"title":"Response Read Attendance Attendance  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/attendance/{attendance_id}":{"get":{"tags":["Attendance"],"summary":"Read Attendance Record","operationId":"read_attendance_record_attendance__attendance_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"attendance_id","in":"path","required":true,"schema":{"type":"integer","title":"Attendance Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Attendance"],"summary":"Update Attendance","operationId":"update_attendance_attendance__attendance_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"attendance_id","in":"path","required":true,"schema":{"type":"integer","title":"Attendance Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Attendance"],"summary":"Delete Attendance","operationId":"delete_attendance_attendance__attendance_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"attendance_id","in":"path","required":true,"schema":{"type":"integer","title":"Attendance Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/leaves/":{"post":{"tags":["leaves"],"summary":"Create Leave","operationId":"create_leave_leaves__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["leaves"],"summary":"Read Leaves","operationId":"read_leaves_leaves__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeaveRequestOut"},"title":"Response Read Leaves Leaves  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/leaves/{leave_id}/approve":{"post":{"tags":["leaves"],"summary":"Approve Leave","operationId":"approve_leave_leaves__leave_id__approve_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"leave_id","in":"path","required":true,"schema":{"type":"integer","title":"Leave Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/leaves/{leave_id}/deny":{"post":{"tags":["leaves"],"summary":"Deny Leave","operationId":"deny_leave_leaves__leave_id__deny_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"leave_id","in":"path","required":true,"schema":{"type":"integer","title":"Leave Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/leaves/{leave_id}":{"get":{"tags":["leaves"],"summary":"Read Leave","operationId":"read_leave_leaves__leave_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"leave_id","in":"path","required":true,"schema":{"type":"integer","title":"Leave Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["leaves"],"summary":"Update Leave","operationId":"update_leave_leaves__leave_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"leave_id","in":"path","required":true,"schema":{"type":"integer","title":"Leave Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["leaves"],"summary":"Delete Leave","operationId":"delete_leave_leaves__leave_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"leave_id","in":"path","required":true,"schema":{"type":"integer","title":"Leave Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/register":{"post":{"summary":"Register","operationId":"register_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"summary":"Read Users Me","operationId":"read_users_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"AttendanceCreate":{"properties":{"employee_id":{"type":"integer","title":"Employee Id"},"login_time":{"type":"string","format":"date-time","title":"Login Time"},"logout_time":{"type":"string","format":"date-time","title":"Logout Time"},"on_leave":{"type":"boolean","title":"On Leave","default":false},"work_hours":{"type":"number","title":"Work Hours"}},"type":"object","required":["employee_id","login_time"],"title":"AttendanceCreate"},"AttendanceOut":{"properties":{"id":{"type":"integer","title":"Id"},"employee_id":{"type":"integer","title":"Employee Id"},"login_time":{"type":"string","format":"date-time","title":"Login Time"},"logout_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Logout Time"},"on_leave":{"type":"boolean","title":"On Leave"},"work_hours":{"type":"number","title":"Work Hours"},"work_duration":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Work Duration","readOnly":true}},"type":"object","required":["id","employee_id","login_time","logout_time","on_leave","work_hours","work_duration"],"title":"AttendanceOut"},"AttendanceUpdate":{"properties":{"login_time":{"type":"string","format":"date-time","title":"Login Time"},"logout_time":{"type":"string","format":"date-time","title":"Logout Time"},"on_leave":{"type":"boolean","title":"On Leave"},"work_hours":{"type":"integer","title":"Work Hours"}},"type":"object","title":"AttendanceUpdate"},"Body_login_for_access_token_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_token_post"},"EmployeeCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"EmployeeCreate"},"EmployeeOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["id","name","user_id"],"title":"EmployeeOut"},"EmployeeUpdate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","title":"EmployeeUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LeaveRequestCreate":{"properties":{"employee_id":{"type":"integer","title":"Employee Id"},"start_date":{"type":"string","format":"date-time","title":"Start Date"},"end_date":{"type":"string","format":"date-time","title":"End Date"},"leave_type":{"type":"string","title":"Leave Type"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["employee_id","start_date","end_date","leave_type"],"title":"LeaveRequestCreate"},"LeaveRequestOut":{"properties":{"id":{"type":"integer","title":"Id"},"employee_id":{"type":"integer","title":"Employee Id"},"start_date":{"type":"string","format":"date-time","title":"Start Date"},"end_date":{"type":"string","format":"date-time","title":"End Date"},"leave_type":{"type":"string","title":"Leave Type"},"status":{"type":"string","title":"Status"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"}},"type":"object","required":["id","employee_id","start_date","end_date","leave_type","status","reason"],"title":"LeaveRequestOut"},"LeaveRequestUpdate":{"properties":{"start_date":{"type":"string","format":"date-time","title":"Start Date"},"end_date":{"type":"string","format":"date-time","title":"End Date"},"leave_type":{"type":"string","title":"Leave Type"},"status":{"type":"string","title":"Status"},"reason":{"type":"string","title":"Reason"}},"type":"object","title":"LeaveRequestUpdate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role","default":"employee"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"role":{"type":"string","title":"Role"}},"type":"object","required":["id","username","role"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}